Fixed visuals not updating with new location values
