+++ ProxiPalTest_client_20231205232005/app/src/main/java/com/mongodb/app/presentation/compassscreen/CompassViewModel.kt
@@ -9,6 +9,8 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.savedstate.SavedStateRegistryOwner
 import com.mongodb.app.data.SyncRepository
+import com.mongodb.app.data.compassscreen.KM_PER_ONE_LATITUDE_DIFF
+import com.mongodb.app.data.compassscreen.KM_PER_ONE_LONGITUDE_DIFF
 import com.mongodb.app.data.compassscreen.UserLocation
 import com.mongodb.app.presentation.userprofiles.UserProfileViewModel
 import com.mongodb.app.ui.compassscreen.CompassUiState
@@ -16,6 +18,9 @@ import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
+import kotlin.math.atan2
+import kotlin.math.pow
+import kotlin.math.sqrt
 
 class CompassViewModel constructor(
     private var repository: SyncRepository
@@ -49,6 +54,10 @@ class CompassViewModel constructor(
         UserLocation(0.0, 0.0)
     )
 
+    private val _bearing: MutableState<Double> = mutableStateOf(0.0)
+
+    private val _distance: MutableState<Double> = mutableStateOf(0.0)
+
     private val _isMeetingWithMatch: MutableState<Boolean> = mutableStateOf(true)
 
 
@@ -74,6 +83,12 @@ class CompassViewModel constructor(
     // TODO Replace this with the actual location data
     val matchedUserLocation: State<UserLocation>
         get() = _matchedUserLocation
+
+    val bearing: State<Double>
+        get() = _bearing
+
+    val distance: State<Double>
+        get() = _distance
 
     val isMeetingWithMatch: State<Boolean>
         get() = _isMeetingWithMatch
@@ -131,6 +146,7 @@ class CompassViewModel constructor(
         _currentUserLatitude.value = newLatitude
         if (newLatitude.toDoubleOrNull() != null && isValidLatitude(newLatitude.toDouble())){
             _currentUserLocation.value.latitude = newLatitude.toDouble()
+            updateMeasurements()
         }
     }
 
@@ -141,6 +157,7 @@ class CompassViewModel constructor(
         _currentUserLongitude.value = newLongitude
         if (newLongitude.toDoubleOrNull() != null && isValidLongitude(newLongitude.toDouble())){
             _currentUserLocation.value.longitude = newLongitude.toDouble()
+            updateMeasurements()
         }
     }
 
@@ -151,6 +168,7 @@ class CompassViewModel constructor(
         _matchedUserLatitude.value = newLatitude
         if (newLatitude.toDoubleOrNull() != null && isValidLatitude(newLatitude.toDouble())){
             _matchedUserLocation.value.latitude = newLatitude.toDouble()
+            updateMeasurements()
         }
     }
 
@@ -161,7 +179,26 @@ class CompassViewModel constructor(
         _matchedUserLongitude.value = newLongitude
         if (newLongitude.toDoubleOrNull() != null && isValidLongitude(newLongitude.toDouble())){
             _matchedUserLocation.value.longitude = newLongitude.toDouble()
+            updateMeasurements()
         }
+    }
+
+    /**
+     * Updates both the bearing and distance measurements
+     */
+    fun updateMeasurements(){
+        _bearing.value = calculateBearingBetweenPoints(
+            startLatitude = currentUserLocation.value.latitude,
+            startLongitude = currentUserLocation.value.longitude,
+            endLatitude = matchedUserLocation.value.latitude,
+            endLongitude = matchedUserLocation.value.longitude
+        )
+        _distance.value = calculateDistanceBetweenPoints(
+            startLatitude = currentUserLocation.value.latitude,
+            startLongitude = currentUserLocation.value.longitude,
+            endLatitude = matchedUserLocation.value.latitude,
+            endLongitude = matchedUserLocation.value.longitude
+        )
     }
 
     /**
@@ -169,5 +206,56 @@ class CompassViewModel constructor(
      */
     fun toggleMeetingWithMatch(){
         _isMeetingWithMatch.value = !_isMeetingWithMatch.value
+    }
+
+    /**
+     * Calculates the bearing angle, in degrees, between two points
+     * (Bearing angle should be 0 degrees in the +y direction and increase clockwise)
+     */
+    private fun calculateBearingBetweenPoints(
+        startLatitude: Double,
+        startLongitude: Double,
+        endLatitude: Double,
+        endLongitude: Double
+    ): Double {
+        val deltaLatitude = endLatitude - startLatitude
+        val deltaLongitude = endLongitude - startLongitude
+        // The user is at the same location as their match
+        if (deltaLatitude == 0.0 && deltaLongitude == 0.0) {
+            return 0.0
+        }
+        var theta = atan2(deltaLatitude, deltaLongitude)
+        // Convert the angle to degrees
+        theta = Math.toDegrees(theta)
+        // Subtract 90 to make 0 degrees point north instead of east
+        theta -= 90
+        // Make the theta non-negative
+        if (theta < 0) {
+            theta += 360
+        }
+        // Reverse the direction the bearing increases, from counter-clockwise to clockwise
+        theta = 360 - theta
+        // Convert 360 degrees to 0
+        // Theta should never be above 360, but >= check is used anyway
+        if (theta >= 360) {
+            theta -= 360
+        }
+        return theta
+    }
+
+    /**
+     * Calculates the distance, in km, between two points
+     */
+    private fun calculateDistanceBetweenPoints(
+        startLatitude: Double,
+        startLongitude: Double,
+        endLatitude: Double,
+        endLongitude: Double
+    ): Double {
+        // Using the distance formula
+        // Make sure to take into account the actual distance between points
+        val deltaLatitude = (endLatitude - startLatitude) * KM_PER_ONE_LATITUDE_DIFF
+        val deltaLongitude = (endLongitude - startLongitude) * KM_PER_ONE_LONGITUDE_DIFF
+        return sqrt(deltaLatitude.pow(2) + deltaLongitude.pow(2))
     }
 }
